fix(rest): Wrap net/url connection reset EOF errors in user-readable error
For some reason the stdlib http client reports an unexpected closed connection as an io.EOF with no
further information.
This results in connection problems (like an internal firewall/proxy/connection config blocking
connections to a Dynatrace SaaS URL) to just be reported as EOF to users.

This fix makes some assumptions about the internals of the http client and may break in the future,
but if the client is changed to no longer return io.EOF errors it will hopefully return a more
helpful error instead.

see also: https://github.com/golang/go/issues/53472