feat(api): Introduce ListResponse and function for decoding all contained objects
Similar to the existing api.Response an api.ListResponse is added and used in bucket.ListResponse.
For convenience a generic DecodeJSONObjects function is added to allow decoding the contained list
of JSON into struct objects.
To make this easily usable the function is generic, just requiring a caller to provide the desired
type, instead of a prefilled slice with objects of the type to write into.
Sadly methods can not be generic, so unlike DecodeJSON the list version can not be a method of the
ListResponse struct.