refactor: Replace logger struct fields with logger from context
As one of the main users of this library monaco relies on structured logs with slightly
different fields for each call to the api clients, setting a logger only once when constructing
a client is not practical.

Instead of allowing to set the logger but exporting it so users could manually edit it when needed,
we opt for a pattern of viewing logging as an optional sideffect and transporting loggers inside a context.
Using logr functions we attempt to read the logger from the context and if none exists simply discard all
log calls.