fix(buckets): Only attempt update when upserting if create failed with a conflict
The API responds with a 409 Conflict when trying to create a bucket that already exists,
only in this case do we want to attempt an update.
In other cases like 400 Bad Request errors or missing permissions there is no sense in
attempting an update.

Additionally duplication between Upsert and Update is removed, with Upsert calling Update
instead of re-implementing the logic of retrying the getAndUpdate operation.

Also an 'Update' test that wrongly used upsert is fixed to call the intended method.