feat(buckets): Do not update buckets if they are unmodified
As any API update to a bucket results in at least one version increase,
updating a bucket that has not actually changed results in unecessary
operations and inflated version numbers.

To not cause this, the payload is compared to the current configuration
returned by GET and if nothing changed, not HTTP PUT is executed.

To signify that nothing changed a internal sentinel error is introduced,
but hidden from users of exported methods with just return a constructed
statuscode 200 response so users do not have to differentiate between an
actual succesful HTTP PUT update and 'updating' configuration to a state
it already has.