test(buckets): Remove potentially flaky timeout test
Testing whether the Update methods' retries honor timeout passed
via the context is timing based and potentially flaky - we have
already seen it fail in CI when running on windows, but work on
other OSs.

The test is removed, as there is no real value in testing with an assured immediate
timeout, as that would trigger a failure in the initial http
GET call (as http requests of course honor context timeouts)
instead of testing our retry logic.

In the worst case, the http requsts we make would honor the context timeout anyway.