fix(automation): Correctly handle admin access on all API calls
Previous code was missing the adminAccess query parameter in some key requests
like List - this fixed, and to be safe the parameter is intially sent for all
requests, even POST requests which strangely do not document it. (If it is
currently actually possible to create new workflows for other users without
admin permissions, that is very likely to be fixed in the future.)

The logic of setting the query parameter is also simplified. Code already
relied on APIs that do not use it, to ignore it - as 'false' was sent to
all APIs that are not '/workflows'.
Thus rather than sending either 'true' or 'false', we can send 'true' to
all APIs - and only handly a HTTP Forbidden as signal to retry without
adminAccess if it happened on for workflows.