fix: make document::Create more resilient if subsequent patch call receives 404
We call PATCH immediately after POST when creating a document in order to make it public.
However, it can be the case that the subsequent call to the PATCH endpoint happens to early. In this situation the API returns 404 and the call to Create fails. To make this more resilient, a retry mechanism was added.

Signed-off-by: Bernd Warmuth <bernd.warmuth@dynatrace.com>