feat: Allow client factory to create REST clients (#114)
This PR extends the clients factory with the ability to create REST
clients directly, but more importantly, clients that can target classic
APIs (with access token). It also adapts the creation of accounts
clients to be more consistent.

It also adds a helper function `AsResponseOrError(...)`, which should be
used to consistently read and close the response body and allow for
easier error handling.

Calling this returns EITHER a `Response` with the read body and a
successful status OR an error, which could be an `APIError` if a
response was actually received. This can simplify e.g. error handling
for skipping 404s during deletion to something like:
```
_, err = coreapi.AsResponseOrError(d.classicClient.DELETE(ctx, parsedURL.String(), corerest.RequestOptions{}))
	if err != nil {
		apiError := coreapi.APIError{}
		if errors.As(err, &apiError) && apiError.StatusCode == http.StatusNotFound {
			log.Debug("No config with id '%s' found to delete (HTTP 404 response)", id)
			return nil
		}
		return err
	}
```