fix: Acquire lock once outside of sendWithRetries (#134)
In
https://github.com/Dynatrace/dynatrace-configuration-as-code-core/blob/main/api/rest/client.go,
`func (c *Client) sendWithRetries(...)` calls itself recursively, but
(tries to) acquire a lock each time. This causes the function to
deadlock.

Proposed solution is to only acquire this lock once, outside of
`sendWithRetries(...)`.

Only issue is that the rate limiter may sleep once it acquires the lock,
but as this is shared by all goroutines accessing the client, this
should have no impact.