chore: Update Account Management client (#185)
This PR updates the Account Management OpenAPI spec and uses the
openapi-generator to regenerate the client. This results in the
availability of a new method for updating service users.

Two manual fixes are still required, for the service user pagination and
binding boundaries; in the future these might not be needed.