fix(account): Revert `page` and `pageSize` to `integer` for get service users: (#205)
This PR reverts `page` and `pageSize` to `integer` for GET
`/iam/v1/accounts/{accountUuid}/service-users` in the fixed spec and
includes the re-generated code for service user management.

While using `int` results in nicer generated code, it does not conform
to the OpenAPI specification. Instead, ideally the openapi-generator
should map `integer` to `int` for golang if not other constraints are
present. This is requested in
https://github.com/OpenAPITools/openapi-generator/issues/3055, but has
not been implemented yet.