feat(clients): Improved error handling
Instead of 4 different error struct (`ErrorPermissionGet`, `ErrorPermissionCreate`, `ErrorPermissionUpdate`, and `ErrorPermissionDelete`), there is only one struct (`ErrorPermissions`) which holds an instance of `ClientOperation`, which can be either `GET`, `POST`, `PUT`, or `DELETE`. `Unwrap()` is implemented and `Error()` is improved, s.t. it also prints the reader-friendly version of the client operation.