ci: Add e2e tests
This change introduces e2e tests for the core library.
E2E tests are executed on the main branch for every change, as well as every day at midnight.

Additionally, e2e tests can be 'opt-in' using labels. This ensures that we don't 'just' execute the tests on each pull request without reviewing, but that we need to actively review it before setting the label.